[Test] Separated scenarios from base tests, moved helpers to another file
